backend/requirements.txt
Flask-SQLAlchemy
Flask-Migrate
requests
python-dotenv
pg8000
numpy
pandas
scikit-learn
joblib
matplotlib
seaborn
SQLAlchemy